% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCell_aggDA.R
\name{aggDA}
\alias{aggDA}
\title{calculate the number of cells
\code{aggDA} calculate number of cells per node-sample combination.}
\usage{
aggDA(d_sce, tree, sample_id = "sample_id", group_id = "group_id",
  cluster_id = "cluster_id", FUN = sum)
}
\arguments{
\item{d_sce}{A \code{SingleCellExperiment} object}

\item{tree}{A phylo tree.}

\item{sample_id}{The column name of sample id.}

\item{group_id}{The column name of group id.}

\item{cluster_id}{The column name of clusters.}

\item{FUN}{The function}
}
\value{
A TreeSummarizedExperiment object. In matrix of \code{assays},
  samples in columns and clusters (or nodes) in rows.
}
\description{
calculate the number of cells
\code{aggDA} calculate number of cells per node-sample combination.
}
\examples{

}
