% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggData.R
\name{aggData}
\alias{aggData}
\title{Perform data aggregations on the tree structure}
\usage{
aggData(x, onRow = TRUE, FUN = sum, message = FALSE)
}
\arguments{
\item{x}{A TreeSummarizedExperiment object.}

\item{onRow}{A logical value, TRUE or FALSE.}

\item{FUN}{A function for aggregation.}

\item{message}{A logical value, TRUE or FALSE.}
}
\value{
A TreeSummarizedExperiment object
}
\description{
\code{aggData} aggregates data on each node of the tree. Users could decide
on which dimension (row or column) and how should the aggregation be
performed.
}
\examples{
#' # assays data
set.seed(1)
toyTable <- matrix(rnbinom(20, size = 1, mu = 10), nrow = 5)
colnames(toyTable) <- paste(rep(LETTERS[1:2], each = 2),
                            rep(1:2, 2), sep = "_")
rownames(toyTable) <- paste("entity", seq_len(5), sep = "")

toyTable

# the column data
colInf <- DataFrame(gg = c(1, 2, 3, 3),
                    group = rep(LETTERS[1:2], each = 2),
                    row.names = colnames(toyTable))
colInf

# the toy tree
library(ape)
set.seed(4)
treeC <- rtree(4)
treeC$node.label <- c("All", "GroupA", "GroupB")

library(ggtree)
ggtree(treeC, size = 2) +
    geom_text2(aes(label = node), color = "darkblue",
               hjust = -0.5, vjust = 0.7, size = 6) +
    geom_text2(aes(label = label), color = "darkorange",
               hjust = -0.1, vjust = -0.7, size = 6)


tse <- TreeSummarizedExperiment(assays = list(toyTable),
                                colData = colInf,
                                colTree = treeC,
                                colNodeLab = treeC$tip.label)

aggCol <- aggData(x = tse, onRow = FALSE, FUN = sum)

}
\author{
Ruizhu HUANG
}
