% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggDS.R
\name{aggDS}
\alias{aggDS}
\title{aggregate data
\code{aggDS} aggregate data based on the tree.}
\usage{
aggDS(TSE, assay = "counts", sample_id = "sample_id",
  group_id = "group_id", cluster_id = "cluster_id", FUN = sum,
  message = FALSE)
}
\arguments{
\item{TSE}{A \code{TreeSummarizedExperiment} object}

\item{assay}{A name or a numeric value to index a matrix in \code{assays}.}

\item{sample_id}{The column name of sample id.}

\item{group_id}{The column name of group id.}

\item{cluster_id}{The column name of clusters.}

\item{FUN}{The function}

\item{message}{TRUE or FALSE. If TRUE, the running process is shown.}
}
\value{
A singleCellExperiment object. In \code{assays}, each matrix is data of a node on the tree. Rows of matrix represent genes and columns represent samples.
}
\description{
aggregate data
\code{aggDS} aggregate data based on the tree.
}
\examples{

}
