% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgScore.R
\name{avgScore}
\alias{avgScore}
\title{Generate average score in a family}
\usage{
avgScore(tree, scoreData, nodeCol, scoreCol)
}
\arguments{
\item{tree}{A phylo object}

\item{scoreData}{A data frame that includes at least two columns. One column
stores the number of the node, and the other stores the original score of
the corresponding node.}

\item{nodeCol}{The name of the column that stores the number of the node.}

\item{scoreCol}{The name of the column that stores the original score of the
node.}
}
\description{
\code{avgScore} takes the tree structure into account when calculating the
score for an internal node. If an internal node A has two children B and C,
\code{(A->B, A->C)}, the new score at node A would be calculated as the
mean of the scores in the whole family (A, B and C).
}
\examples{

library(TreeSummarizedExperiment)
library(ggtree)
library(dplyr)

# tree
data("tinyTree")
ggtree(tinyTree) + geom_text2(aes(label = node))

# score
exScore <- data.frame(nodeNum = 1:19, score = (1:19)/10)

# update the score based on the family provided by the tree
newScore <- avgScore(tree = tinyTree, scoreData = exScore,
                      nodeCol = "nodeNum", scoreCol = "score")

# visualize the result
df <- newScore \%>\% 
      rename(node = nodeNum) \%>\%    
      mutate(score = round(score, 3)) \%>\%
      mutate(avgScore = round(avgScore, 3))

# the original scores are in black texts and the new ones in blue
ggtree(tinyTree) \%<+\% df + 
geom_text2(aes(label = score), hjust = -0.05) +
geom_text2(aes(label = avgScore, hjust = -1.2), color = "blue") +
xlim(0, 5)  
}
\author{
Ruizhu HUANG
}
