% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCell_buildTree.R
\name{buildTree2}
\alias{buildTree2}
\title{build a tree from clusters 
\code{buildTree2} creates a tree (\code{phylo} class).}
\usage{
buildTree2(d_se, column_cluster, npcs = 10, dist_method = "euclidean",
  hclust_method = "average")
}
\arguments{
\item{d_se}{A
\code{\link[TreeSummarizedExperiment]{TreeSummarizedExperiment}} object
output from \code{\link{prepareData}}.}

\item{column_cluster}{A vector of cluster labels. It should have the same
length as the number of cells in \code{data_HVG}.}

\item{npcs}{A numeric value. The number of principal components used to
calculated distances among clusters. Default (NULL) takes the minumum value
of the number of clusters and the number of highly variable genes.}

\item{dist_method}{the distance measure to be used. This must be one of
"euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".
Any unambiguous substring can be given. Please refer to \code{method} in
\code{\link[stats]{dist}}}

\item{hclust_method}{the agglomeration method to be used. This should be (an
unambiguous abbreviation of) one of "ward.D", "ward.D2", "single",
"complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC)
or "centroid" (= UPGMC). Please refer to \code{method} in
\code{\link[stats]{hclust}}}
}
\value{
A phylo object
}
\description{
Distances among clusters are calculated based on the median values of
features (genes) of clusters in reduced dimensions (PCA)
}
\examples{

library(ggtree)
# data
ncell <- 500
ngene <- 20
dat <- matrix(rpois(ngene*ncell, lambda = 20), nrow = ngene)

# clusters
clus <- sample(1:50, 500, replace = TRUE)

# tree
tr <- buildTree2(data_HVG = dat, cluster_id = clus)

ggtree(tr, branch.length = "none") +
   geom_tiplab(size = 2) +
   xlim(0, 20)
}
