% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeTree.R
\name{changeTree}
\alias{changeTree}
\title{change the tree in TreeSummarizedExperiment}
\usage{
changeTree(tse, new_colTree = NULL, new_rowTree = NULL,
  lab_colTree = NULL, lab_rowTree = NULL)
}
\arguments{
\item{tse}{A TreeSummarizedExperiment object.}

\item{new_colTree}{A phylo object. A new tree to replace the current column
tree.}

\item{new_rowTree}{A phylo object. A new tree to replace the current row
tree.}

\item{lab_colTree}{A data frame. A data frame with two columns (oldLab and
newLab), one having the node labels of the current column tree, and the
other having those of the new tree.}

\item{lab_rowTree}{A data frame. A data frame with two columns (oldLab and
newLab), one having the node labels of the current row tree, and the other
having those of the new tree.}
}
\value{
A TreeSummarizedExperiment object
}
\description{
\code{changeTree} changes a row or column tree to a new one by matching the
node labels.
}
