% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectData.R
\name{collectData}
\alias{collectData}
\title{collect data for each node of the tree}
\usage{
collectData(x, rowLevel = NULL, colLevel = NULL, message = FALSE)
}
\arguments{
\item{x}{A TreeSummarizedExperiment object.}

\item{message}{A logcical value, TRUE or FALSE.}

\item{onRow}{A logical value, TRUE or FALSE.}
}
\value{
A TreeSummarizedExperiment objects
}
\description{
\code{collectData} generates data for each node of the tree by collecting
data on its descendant leaves. For leaf node, the collected data is from
itself. Users could decide on which dimension (row or column) the collection
should perform.
}
\examples{
#' 
library(ape)
library(ggtree)
set.seed(1)

# tree
exTree <- rtree(5)
ggtree(exTree) + geom_text2(aes(label = node), hjust = -0.5)

# toy table
toyTable <- matrix(1:25, nrow = 5)
rownames(toyTable) <- exTree$tip.label

# row data
rowD <- DataFrame(letters = sample(letters[1:2], 5, replace = TRUE))
# tse
lse <- TreeSummarizedExperiment(assays = list(toyTable),
                                rowTree = exTree,
                                rowData = rowD)

# collect data for each node
tse <- collectData(x = lse)



}
\author{
Ruizhu HUANG
}
