% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDf.R
\name{convertDf}
\alias{convertDf}
\title{convert a TreeSummarizedExperiment object to a data frame}
\usage{
convertDf(tse, onRow = TRUE, assayNum = NULL)
}
\arguments{
\item{tse}{A TreeSummarizedExperiment object.}

\item{onRow}{A logical value, TRUE or FALSE. If entities are in rows and
samples are in columns, then TRUE is required; if the other way around,
then FALSE is required.}

\item{assayNum}{A numeric value to specify which table in the \code{assays}
slot should be used. The first table is used by default.}
}
\value{
A data frame
}
\description{
\code{convertDf} converts a TreeSummarizedExperiment object to a data frame.
The data frame includes observed values of entities on each sample and the
sample information that is originally stored in the column data (if samples
are in the column dimension of \code{assays} table) or the row data (if
samples are in the row dimension of \code{assays} table).
}
\examples{
  

library(ape)
library(TreeSummarizedExperiment)
library(ggtree)

set.seed(1)
treeR <- rtree(5)


ggtree(treeR, size = 2) +
    geom_text2(aes(label = node), color = "darkblue",
               hjust = -0.5, vjust = 0.7, size = 6) +
    geom_text2(aes(label = label), color = "darkorange",
               hjust = -0.1, vjust = -0.7, size = 6)
               
toyTable <- matrix(rnbinom(20, size = 1, mu = 10), nrow = 5)
colnames(toyTable) <- paste(rep(LETTERS[1:2], each = 2),
                            rep(1:2, 2), sep = "_")
rownames(toyTable) <- treeR$tip.label

toyTable

# the column data
colInf <- DataFrame(gg = c(1, 2, 3, 3),
                    gr = rep(LETTERS[1:2], each = 2),
                    row.names = colnames(toyTable))
colInf

# the toy tree


tse <- TreeSummarizedExperiment(assays = list(toyTable),
                                colData = colInf,
                                rowTree = treeR)

df <- convertDf(tse = tse, onRow = TRUE)
 
}
\author{
Ruizhu Huang
}
