% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvGLM.R
\name{cvGLM}
\alias{cvGLM}
\title{k-fold cross validation for logistic (fixed or mixed-effects) model}
\usage{
cvGLM(tse, assayNum = NULL, onRow = TRUE, kFold = NULL,
  confounder = NULL, respY = "group", lmeTip = FALSE,
  lmeInNode = TRUE, familyTip = "binomial",
  familyInNode = "binomial", message = FALSE)
}
\arguments{
\item{tse}{A TreeSummarizedExperiment object}

\item{assayNum}{A numeric vector. It specifies which matrix-like elements
in assays will be used to do analysis. If NULL, the first one is used.}

\item{onRow}{TRUE or FALSE. If TRUE, entities are on rows and samples on
columns.}

\item{kFold}{A numeric value to specify k-fold cross validation.}

\item{confounder}{The confounder variables. The names of columns in
\code{colData} if \code{onRow=TRUE} that store the confounders.}

\item{respY}{The response variables. The names of columns in
\code{colData} if \code{onRow=TRUE} that store the response variable.}

\item{lmeTip}{TRUE or FALSE. Default is FALSE. If TRUE, logistic
mixed-effects model is used on the leaf nodes.}

\item{lmeInNode}{TRUE or FALSE. Default is FALSE. If TRUE, logistic
mixed-effects model is used on the internal nodes.}

\item{familyTip}{"binomial" or "quasibinomial".}

\item{familyInNode}{"binomial" or "quasibinomial"}

\item{message}{TRUE or FALSE}
}
\value{
a data frame
}
\description{
\code{runGLM} is to run k-fold cross validation for logistic (fixed or
mixed-effects) model.
}
\examples{
library(ape)
library(ggtree)

n <- 20
nSam <- c(50, 50)
lib <- 500
set.seed(3)
exTree <- rtree(n)
ggtree(exTree) + geom_text2(aes(label = node))
prob <- rep(1/n, n)
names(prob) <- exTree$tip.label
lse <- simMult(pr = prob, libSize = lib, 
               tree = exTree,
               minTip.A = 3, maxTip.A = 6,
               nSam = nSam, scenario = "S1")

tse2 <- aggData(x = lse, onRow = TRUE)
out2 <- cvGLM(tse = tse2, onRow = TRUE, kFold = 10,
              lmeTip = FALSE, lmeInNode = FALSE,
              familyTip = "quasibinomial", 
              familyInNode = "quasibinomial")

}
\author{
Ruizhu Huang
}
