% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCell_dsSC.R
\name{.designMatrix}
\alias{.designMatrix}
\title{Create design matrix}
\usage{
.designMatrix(data, cols = NULL)
}
\arguments{
\item{data}{A \code{data.frame} or \code{DataFrame}.}

\item{cols}{A numeric vector. Specify columns to include in the design of
model matrix. Default is to include all columns.}
}
\value{
a matrix.
}
\description{
\code{.designMatrix} creates a design matrix by expanding factors to a set of
dummay variables and epanding interactions similarly.
}
\details{
\code{.designMatrix} creates a design matrix using the data extracted from
\code{data}. \code{cols} specifies the columns to extract.
\code{.designMatrix} is built on \code{\link[stats]{model.matrix}} with
\code{contrasts.arg = NULL}.
}
\keyword{internal}
