% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{.doData}
\alias{.doData}
\title{Simulate a count table}
\usage{
.doData(
  tree = NULL,
  data = NULL,
  scenario = "S1",
  from.A = NULL,
  from.B = NULL,
  minTip.A = 0,
  maxTip.A = Inf,
  minTip.B = 0,
  maxTip.B = Inf,
  minPr.A = 0,
  maxPr.A = 1,
  ratio = 2,
  adjB = NULL,
  pct = 0.6,
  nSam = c(50, 50),
  mu = 50,
  size = 10000,
  n = 1,
  FUN = sum
)
}
\arguments{
\item{tree}{A phylo object}

\item{data}{A matrix, representing a count table from real data. It has the
entities corresponding to tree leaves in the row and samples in the column.}

\item{scenario}{\dQuote{S1}, \dQuote{S2}, or \dQuote{S3} (see
\bold{Details}). Default is \dQuote{S1}.}

\item{from.A, from.B}{The branch node labels of branches A and B for which the
signal is swapped. Default, both are NULL. In simulation, we select two
branches (A & B) to have differential abundance under different conditions.
One could specify these two branches or let \code{.doData} choose. (Note: If
\code{from.A} is NULL, \code{from.B} is set to NULL).}

\item{minTip.A}{The minimum number of leaves in branch A}

\item{maxTip.A}{The maximum number of leaves in branch A}

\item{minTip.B}{The minimum number of leaves in branch B}

\item{maxTip.B}{The maximum number of leaves in branch B}

\item{minPr.A}{A numeric value selected from 0 to 1. The minimum abundance
proportion of leaves in branch A}

\item{maxPr.A}{A numeric value selected from 0 to 1.The maximum abundance
proportion of leaves in branch A}

\item{ratio}{The proportion ratio of branch B to branch A. This value is used
to select branches(see \bold{Details}). If there are no branches having
exactly this ratio, the pair with the value closest to \code{ratio} would
be selected.}

\item{adjB}{a numeric value selected from 0 and 1 (only for \code{scenario}
is \dQuote{S3}). Default is NULL. If NULL, branch A and branch B swap their
proportions. If a numeric value, e.g. 0.1, then branch B decreases to its
one tenth proportion and the decrease in branch B is added to branch A. For
example, assume there are two experimental conditions (C1 & C2), branch A
has 10 and branch B has 40 in C1. If adjB is set to 0.1, then in C2 branch
B becomes 4 and branch A 46 so that the total proportion stays the same.}

\item{pct}{a numeric value selected from 0 and 1. The percentage of leaves in
branch B that have differential abundance under different conditions (only
for scenario \dQuote{S3})}

\item{nSam}{A numeric vector of length 2, containing the sample size for two
different conditions}

\item{mu, size}{The parameters of the Negative Binomial distribution. (see mu
and size in \code{\link[stats:NegBinomial]{rnbinom}}). Parameters used to
generate the library size for each simulated sample.}

\item{n}{A numeric value to specify how many count tables would be generated
with the same settings. Default is one and one count table would be
obtained at the end. If above one, the output of \code{.doData} is a list of
matrices (count tables). This is useful, when one needs multiple
simulations.}

\item{FUN}{A function to derive the count at each internal node based on its
descendant leaves, e.g. sum, mean. The argument of the function is a
numeric vector with the counts of an internal node's descendant leaves.}
}
\value{
a list of objects \item{FC}{the fold change of entities correspondint
  to the tree leaves.} \item{Count}{a list of count table or a count table.
  Entities on the row and samples in the column. Each count table includes
  entities corresponding to all nodes on the tree structure.}
  \item{Branch}{the information about two selected branches.} \describe{
  \item{A}{the branch node label of branch A} \item{B}{the branch node label
  of branch B} \item{ratio}{the count proportion ratio of branch B to branch
  A} \item{A_tips}{the number of leaves on branch A} \item{B_tips}{the number
  of leaves on branch B} \item{A_prop}{the count proportion of branch A (not
  above 1)} \item{B_prop}{the count proportion of branch B (not above 1)} }
}
\description{
\code{.doData} creates a count table for all nodes of a tree under two
different groups such that the tree would have different abundance patterns
in the different conditions.
}
\details{
\code{.doData} simulates a count table for entities which are
  corresponding to the nodes of a tree. The entities are in rows and the
  samples from different groups or conditions are in columns. The library
  size of each sample is sampled from a Negative Binomial distribution with
  mean and size specified by the arguments \code{mu} and \code{size}. The
  counts of entities, which are located on the tree leaves, in the same
  sample are assumed to follow a Dirichlet-Multinomial distribution. The
  parameters for the Dirichlet-Multinomial distribution are estimated from a
  real data set specified by the argument \code{data} via the function
  \code{dirmult} (see \code{\link[dirmult]{dirmult}}). To generate different
  abundance patterns under different conditions, we provide three different
  scenarios, \dQuote{S1}, \dQuote{S2}, and \dQuote{S3} (specified via
  \code{scenario}). \itemize{ \item S1: two branches are selected to swap
  their proportions, and leaves on the same branch have the same fold change.
  \item S2: two branches are selected to swap their proportions. Leaves in
  the same branch have different fold changes but same direction (either
  increase or decrease). \item S3: two branches are selected. One branch has
  its proportion swapped with the proportion of some leaves from the other
  branch.}
}
\examples{
\dontrun{
if(require(GUniFrac)){
data("throat.otu.tab")
data("throat.tree")

dat <- .doData(tree = throat.tree,
data = as.matrix(t(throat.otu.tab)),
ratio = 2)
}
}
}
\author{
Ruizhu Huang
}
\keyword{internal}
