% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{.doFC}
\alias{.doFC}
\title{Generate the fold change}
\usage{
.doFC(
  tree = NULL,
  data = NULL,
  scenario = "BS",
  branchA = NULL,
  branchB = NULL,
  ratio = 1,
  adjB = NULL,
  pct = 1
)
}
\arguments{
\item{tree}{A phylo object}

\item{data}{The real data (count table)}

\item{scenario}{Scenarios (\dQuote{BS}, \dQuote{US}, \dQuote{SS})}

\item{branchA}{The branch node label of branch A.}

\item{branchB}{The branh node label of branch B.}

\item{ratio}{The proportion ratio between \code{branchB} and \code{branchA}
(B/A)}

\item{adjB}{A numeric value between 0 and 1 (only for \code{scenario}
is \dQuote{SS}). Default is NULL. If NULL, branch A and branch B swap their
proportions. If a numeric value, e.g. 0.1, then branch B decreases to its
one tenth proportion and the decrease in branch B is added to branch A.
For example, assume there are two experimental conditions (C1 & C2), branch
A has 10 and branch B has 40 in C1. If adjB is set to 0.1, then in C2 branch
B becomes 4 and branch A 46 so that the total proportion stays the same.}

\item{pct}{The percentage (in number) of the leaves in branchA that will
swap with branchB. The default is 1.}
}
\value{
numeric vector
}
\description{
\code{.doFC} generates fold changes for different scenarios
}
\author{
Ruizhu Huang
}
\keyword{internal}
