% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{.fdr0}
\alias{.fdr0}
\title{Calculate true positives and positives}
\usage{
.fdr0(tree, truth = NULL, found = NULL, only.leaf = TRUE)
}
\arguments{
\item{tree}{A phylo object}

\item{truth}{Nodes that have signals (eg. differentally abundant at different
experimental conditions.).}

\item{found}{Nodes that have been found to have signal}

\item{only.leaf}{A logical value, TRUE or FALSE. If TRUE, true positive rate
is calculated at the leaf (tip) level; otherwise it is calculated at node
level. The default is TRUE}
}
\value{
a vector
}
\description{
\code{.fdr0} calculates the number of false discovery and the total number of
discovery on a tree structure at leaf or node level.
}
\keyword{internal}
