% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{.infLoc}
\alias{.infLoc}
\title{Provide the information of two branches}
\usage{
.infLoc(tree = NULL, data = NULL, from.A = NULL, from.B = NULL)
}
\arguments{
\item{tree}{A phylo object}

\item{data}{A count table (a matrix or a data frame). It has tree leaves in
rows and samples from different conditions in  columns.}

\item{from.A, from.B}{The branch node labels of Branch A, B.}
}
\value{
A data frame of one row
}
\description{
\code{.infLoc} is to give information of two branches about the count
proportion and the number of leaves
}
\author{
Ruizhu Huang
}
\keyword{internal}
