% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{.pickLoc}
\alias{.pickLoc}
\title{Select branches}
\usage{
.pickLoc(
  tree = NULL,
  data = NULL,
  from.A = NULL,
  minTip.A = 0,
  maxTip.A = Inf,
  minTip.B = 0,
  maxTip.B = Inf,
  minPr.A = 0,
  maxPr.A = 1,
  ratio = 1
)
}
\arguments{
\item{tree}{A phylo object}

\item{data}{A count table (a matrix or a data frame). It has tree leaves in
rows and samples from different conditions in columns.}

\item{from.A}{The branch node label of branch A. In simulation, we select two
branches (A & B) to have differential abundance under different conditions.
If from.A is specified, then branch A is fixed. If from.A is NULL, one can
find a suitable branch which meets the criteria specified in \code{minTip.A},
\code{maxTip.A}, \code{minPr.A} and \code{maxPr.A}.}

\item{minTip.A}{The minimum number of leaves in branch A}

\item{maxTip.A}{The maximum number of leaves in branch A}

\item{minTip.B}{The minimum number of leaves in branch B}

\item{maxTip.B}{The maximum number of leaves in branch B}

\item{minPr.A}{The minimum abundance proportion of leaves in branch A}

\item{maxPr.A}{The maximum abundance proportion of leaves in branch A}

\item{ratio}{The proportion ratio of branch B to branch A. This value is
used to select branches(see \bold{Details}). If there are no branches having
exactly this ratio, the pair with the value closest to \code{ratio} would
be selected.}
}
\value{
a data frame of one row
}
\description{
\code{.pickLoc} selects two branches which meet the criteria specified by
the arguments
}
\author{
Ruizhu Huang
}
\keyword{internal}
