% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCand.R
\name{getCand}
\alias{getCand}
\title{search candidates under different thresholds}
\usage{
getCand(
  tree,
  t = NULL,
  score_data,
  node_column,
  p_column,
  sign_column,
  threshold = 0.05,
  pct_NA = 0.5,
  message = FALSE
)
}
\arguments{
\item{tree}{A phylo object.}

\item{t}{A sequence of values with the range between 0 and 1. Thresholds used
to search candidates. The default is to use a sequence \code{c(seq(0, 0.04,
by = 0.01), seq(0.05, 1, by = 0.05))}}

\item{score_data}{A data frame includes at least one column about the nodes,
one column about the p value (\code{p_column}) and one column about the
direction of change (\code{sign_column}).}

\item{node_column}{The name of the column that gives the node information.}

\item{p_column}{The name of the column that gives p values of nodes.}

\item{sign_column}{The name of the column that gives the direction of the
difference.}

\item{threshold}{Default is 0.05. A value to hinder an internal to be picked
due to randomness.}

\item{message}{A logical value, TRUE or FALSE. Default is FALSE. If TRUE, the
message about running process is printed out.}

\item{pct_na}{A numeric value. Default is 0.5. Internal nodes that are 
finally selected should have more than half of direct children (> 0.5) 
without missing p values.}
}
\value{
A list includes \code{candidate_list} and \code{score_data}. 
\item{condidate_list}{A list of candidates under different thresholds.}
\item{score_data}{a data frame that includes columns from the input
    \code{score_data} and additional columns to store score U and s under
    different thresholds.}
}
\description{
\code{getCand} search candidates under different thresholds
}
\examples{
library(TreeSummarizedExperiment)
library(ggtree)

data(tinyTree)
ggtree(tinyTree, branch.length = "none") +
   geom_text2(aes(label = node)) +
   geom_hilight(node = 13, fill = "blue", alpha = 0.5) +
   geom_hilight(node = 18, fill = "orange", alpha = 0.5)
set.seed(2)
pv <- runif(19, 0, 1)
pv[c(1:5, 13, 14, 18)] <- runif(8, 0, 0.001)

fc <- sample(c(-1, 1), 19, replace = TRUE)
fc[c(1:3, 13, 14)] <- 1
fc[c(4, 5, 18)] <- -1
df <- data.frame(node = 1:19,
                 pvalue = pv,
                 foldChange = fc)

ll <- getCand(tree = tinyTree, score_data = df,
              node_column = "node", 
              p_column = "pvalue", 
              sign_column = "foldChange")
}
\author{
Ruizhu Huang
}
