% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTree.R
\name{getTree}
\alias{getTree}
\title{build a tree from clusters 
\code{getTree} creates a tree (\code{phylo} class).}
\usage{
getTree(SE, column_cluster = "cluster_id", use_marker = NULL,
  npcs = NULL, dist_method = "euclidean", hclust_method = "average")
}
\arguments{
\item{SE}{A
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object
output from \code{\link{medianByClusterMarker}}.}

\item{column_cluster}{The name of the column that stores the cluster id.}

\item{use_marker}{A logical vector that has the same length as the number of
columns in \code{x} to specify markers to build the tree.}

\item{npcs}{A numeric value. The number of principal components used to
calculated distances among clusters. Default (NULL) takes the minumum value
between the number of clusters and the number of markers.}

\item{dist_method}{the distance measure to be used. This must be one of
"euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".
Any unambiguous substring can be given. Please refer to \code{method} in
\code{\link[stats]{dist}}}

\item{hclust_method}{the agglomeration method to be used. This should be (an
unambiguous abbreviation of) one of "ward.D", "ward.D2", "single",
"complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC)
or "centroid" (= UPGMC). Please refer to \code{method} in
\code{\link[stats]{hclust}}}
}
\value{
A phylo object
}
\description{
Distances among clusters are calculated based on the median values of
features (genes) of clusters in reduced dimensions (PCA)
}
\examples{

library(ggtree)

# data
ncell <- 500
ngene <- 20
dat <- matrix(rpois(ngene*ncell, lambda = 20), ncol = ngene)

# clusters
clus <- sample(letters, 500, replace = TRUE)
rowD <- data.frame(cluster_id = clus,
                   stringsAsFactors = FALSE)
d_se <- SummarizedExperiment(assays = list(dat),
                             rowData = rowD)
dx <- medianByClusterMarker(SE = d_se, marker_in_column = TRUE,
                            column_cluster = "cluster_id")
             
# tree
tr <- getTree(SE = dx, column_cluster = "cluster_id",
              npcs = 20)
# leaf nodes are labelled with cluster id
tr$tip.label
ggtree(tr, branch.length = "none") +
   geom_tiplab(size = 2) +
   xlim(0, 20)
}
