% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianByClusterMarker.R
\name{medianByClusterMarker}
\alias{medianByClusterMarker}
\title{calculate median values of markers for each cluster}
\usage{
medianByClusterMarker(
  SE,
  marker_in_column = TRUE,
  column_cluster = "cluster_id",
  use_marker = NULL
)
}
\arguments{
\item{SE}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}} object.}

\item{marker_in_column}{A logical value, either TRUE or FALSE. The default is
TRUE (markers are in columns)}

\item{column_cluster}{The name of the column that has information about
cluster id.}

\item{use_marker}{A logical vector including FALSE and TRUE to specify
markers to be used. The default is NULL and all markers are used.}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}} object. It
  includes clusters in rows and the median value of each marker in each
  cluster in columns
}
\description{
\code{medianByClusterMarker} calculates median value of each marker in each
cluster. The input data should be a
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object. The markers
could either be in columns or in rows of the \code{assays} slot.
}
\examples{
library(TreeSummarizedExperiment)
set.seed(1)
count <- matrix(rpois(n = 1000, lambda = 10),
                nrow = 100)
colnames(count) <- paste0("mk", 1:10)
rowD <- data.frame("cluster" = sample(1:6, 100, replace = TRUE))
colD <- data.frame(type_marker = rep(c(FALSE, TRUE), each = 5))
d_se <- TreeSummarizedExperiment(assays = list(count),
                                 rowData = rowD,
                                 colData = colD)
d_med <- medianByClusterMarker(SE = d_se, marker_in_column = TRUE,
                               column_cluster = "cluster",
                               use_marker = colData(d_se)$type_marker)
d_med
}
\author{
Ruizhu Huang
}
