% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parEstimate.R
\name{parEstimate}
\alias{parEstimate}
\title{Parameter estimation in Dirichlet-multinomial distribution}
\usage{
parEstimate(obj)
}
\arguments{
\item{obj}{A matrix or leafSummarizedExperiment. Samples in the columns and
entities in the rows.}
}
\value{
A list including \dQuote{pi} and \dQuote{theta}
}
\description{
\code{parEstimate} is a wrapper function generated from the function
\code{\link[dirmult]{dirmult}} with default setting on \code{init},
\code{initscalar}, \code{epsilon}, \code{trace} and \code{mode}. It allows
the input \code{obj} to accept \code{matrix} or
\code{leafSummarizedExperiment} and output the value \code{pi} and
\code{theta}.
}
\examples{


set.seed(1)
y <- matrix(rnbinom(100,size=1,mu=10),nrow=10)
colnames(y) <- paste("S", 1:10, sep = "")
rownames(y) <- tinyTree$tip.label


toy_tse <- TreeSummarizedExperiment(rowTree = tinyTree,
                                    assays = list(y))
res <- parEstimate(obj = toy_tse)

metadata(res)$assays.par

}
\author{
Ruizhu Huang
}
