% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGLM.R
\name{runGLM}
\alias{runGLM}
\title{run logistic (mixed-effects) model}
\usage{
runGLM(tse, onRow = TRUE, confounder = NULL, assayNum = NULL,
  respY = "group", lmeTip = FALSE, lmeInNode = TRUE,
  familyTip = "binomial", familyInNode = "binomial", message = TRUE)
}
\arguments{
\item{tse}{A TreeSummarizedExperiment object}

\item{onRow}{TRUE or FALSE}

\item{confounder}{The confounder variables.}

\item{assayNum}{A numeric vector. It specifies which matrix-like elements
in assays will be used to do analysis. If NULL, the first one is used.}

\item{respY}{The response variables.}

\item{lmeTip}{TRUE or FALSE. Default is FALSE. If TRUE, mixed-effects model
is used on the leaf nodes.}

\item{lmeInNode}{TRUE or FALSE. Default is FALSE. If TRUE, mixed-effects model
is used on the internal nodes.}

\item{familyTip}{"binomial" or "quasibinomial".}

\item{familyInNode}{"binomial" or "quasibinomial"}
}
\value{
a data frame
}
\description{
\code{runGLM} is to run logistic model analysis for hierarchical data. For
data on the leaf level, the logistic fixed-effect model is used. However, for
data on the internal nodes, uses could decide whether the mixed-effects or
the fixed effect model should be used.
}
\examples{
library(ape)
library(ggtree)

n <- 20
nSam <- c(50, 50)
lib <- 500
set.seed(3)
exTree <- rtree(n)
ggtree(exTree) + geom_text2(aes(label = node))
prob <- rep(1/n, n)
names(prob) <- exTree$tip.label
lse <- simMult(pr = prob, libSize = lib, 
               tree = exTree,
               minTip.A = 3, maxTip.A = 6,
               nSam = nSam, scenario = "S1")
tse <- collectData(x = lse, onRow = TRUE)

out <- runGLM(tse = tse, onRow = TRUE)

}
\author{
Ruizhu Huang
}
