% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saliva}
\alias{saliva}
\title{A selected microbial data on human saliva bodysite.}
\format{An object of class \code{TreeSummarizedExperiment} with 1000 rows and 226 columns.}
\usage{
saliva
}
\description{
A TreeSummarizedExperiment object stores the seleted (top abundant) 1000 OTU
from the human microbial data on saliva site from 226 samples.
}
\keyword{datasets}
