% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saliva_2019}
\alias{saliva_2019}
\title{from HMP16SData}
\format{An object of class \code{TreeSummarizedExperiment} with 1156 rows and 156 columns.}
\usage{
saliva_2019
}
\description{
from HMP16SData
}
\keyword{datasets}
