% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saliva_v35}
\alias{saliva_v35}
\title{from HMP16SData}
\format{An object of class \code{TreeSummarizedExperiment} with 956 rows and 153 columns.}
\usage{
saliva_v35
}
\description{
from HMP16SData
}
\keyword{datasets}
