% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreTree.R
\name{scoreTree}
\alias{scoreTree}
\title{Generate score accounting for the family effect}
\usage{
scoreTree(tree, scoreData, nodeCol, scoreCol, FUN)
}
\arguments{
\item{tree}{A phylo object}

\item{scoreData}{A data frame that includes at least two columns. One column
stores the number of the node, and the other stores the original score of
the corresponding node.}

\item{nodeCol}{The name of the column that stores the number of the node.}

\item{scoreCol}{The name of the column that stores the original score of the
node.}

\item{FUN}{A function used to generate the score at each internal node, e.g.
sum.}
}
\description{
\code{scoreTree} takes the tree structure into account when calculating the
score for an internal node. If an internal node A has two children B and C,
\code{(A->B, A->C)}, the new score at node A would be calculated as the mean
(sum or other user-specified function) of the whole family (A, B and C). The
generation starts from the leaves and the new generated scores are used to
update those in higher level of the tree until the root is reached.
}
\examples{

library(TreeSummarizedExperiment)
library(ggtree)
library(dplyr)

# tree
data("tinyTree")
ggtree(tinyTree) + geom_text2(aes(label = node))

# score
exScore <- data.frame(nodeNum = 1:19, score = (1:19)/10)

# update the score based on the family provided by the tree
newScore <- scoreTree(tree = tinyTree, 
                      scoreData = exScore,
                      nodeCol = "nodeNum", 
                      scoreCol = "score",
                      FUN = mean)

# visualize the result
df <- newScore \%>\% 
      rename(node = nodeNum) \%>\%    
      mutate(score = round(score, 3)) \%>\%
      mutate(treeScore = round(treeScore, 3))

# the original scores are in black texts and the new ones in blue
ggtree(tinyTree) \%<+\% df + 
geom_text2(aes(label = score)) +
geom_text2(aes(label = treeScore, hjust = -1), 
color = "blue")   
}
\author{
Ruizhu HUANG
}
