% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchOptimal.R
\name{searchOptimal}
\alias{searchOptimal}
\title{search and output results at the optimal level}
\usage{
searchOptimal(tree, score_data, node_column, p_column, sign_column,
  threshold, score_column = "S", method = "BH", limit_rej = 0.05,
  output_all = TRUE, message = FALSE)
}
\arguments{
\item{tree}{A phylo object.}

\item{score_data}{A data frame includes at least one column about the nodes,
one column about the p value and one column about the difference direction.}

\item{node_column}{The name of the column that gives the node information.}

\item{p_column}{The name of the column that gives p values of nodes.}

\item{sign_column}{The name of the column that gives the direction of the
difference.}

\item{threshold}{A sequence of values with the range between 0 and 1. The
different threshold values that the search is performed at.}

\item{score_column}{The name of the column that will store the generated scores}

\item{method}{The multiple testing correction method. Please refer to the
argument \code{method} in \code{\link[stats]{p.adjust}}}

\item{output_all}{A logical value, TRUE or FALSE. Default is TRUE. If TRUE,
intermediate results are saved.}

\item{message}{A logical value, TRUE or FALSE. Default is FALSE. If TRUE, the
message about running process is printed out.}
}
\value{
a list
\item{results}{
   \itemize{
   \item
   }}
\item{method}{the method used to do multiple testing correction.}
\item{threshold}{the threshold values used to search the levels}
\item{best_threshold}{which threshold value is used to determine the optimal level}
}
\description{
\code{searchOptimal} search the optimal level on the tree to test the NULL
hypotheses and output the result.
}
\examples{
library(TreeSummarizedExperiment)
data(tinyTree)
set.seed(2)
df <- data.frame(node = 1:19, pvalue = runif(19), 
                 foldChange = sample(c(1, -1), 19,
                  replace = TRUE))
                 
out <- searchOptimal(tree = tinyTree, score_data = df,
                     node_column = "node", p_column = "pvalue",
                     sign_column = "foldChange",
                     threshold = seq(0, 1, 0.1))
                      
}
\author{
Ruizhu Huang
}
