% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selNode.R
\name{selNode}
\alias{selNode}
\title{Select branches}
\usage{
selNode(
  pr = NULL,
  obj = NULL,
  data = NULL,
  tree = NULL,
  minTip = 0,
  maxTip = Inf,
  minPr = 0,
  maxPr = 1,
  skip = NULL,
  all = FALSE
)
}
\arguments{
\item{pr}{A named numeric vector to provide proportions of entities.}

\item{obj}{A TreeSummarizedExperiment object.}

\item{data}{A count table with entities in rows and samples in columns.}

\item{tree}{A phylo object}

\item{minTip}{the minimum number of leaves in the selected branch}

\item{maxTip}{The maximum number of leaves in the selected branch}

\item{minPr}{The minimum count proportion of the selected branch in a sample.
A value between 0 and 1.}

\item{maxPr}{The maximum count proportion of the selected branch in a sample.
A value between 0 and 1.}

\item{skip}{A character vector of node labels. These nodes are not the
descendants or the ancestors of the selected branch.}

\item{all}{TRUE or FALSE. Default is FALSE. If FALSE, the branch node of a
branch, which meet the requirements and has the minimum count proportion,
is returned; otherwise branch nodes of all branches meeting the
requirements are returned.}
}
\value{
The node whose descendant branch has the lowest proportion
}
\description{
\code{selNode} selects branches meeting the specified criteria in the number
of leaves and the count proportion.
}
\examples{
set.seed(1)
data("tinyTree")
toyTable <- matrix(rnbinom(40, size = 1, mu = 10), nrow = 10)
colnames(toyTable) <- paste(rep(LETTERS[1:2], each = 2),
rep(1:2, 2), sep = "_")
rownames(toyTable) <- tinyTree$tip.label


dat <- parEstimate(obj = toyTable)
(out1 <- selNode(tree = tinyTree, data = dat, all = TRUE))
(out2 <- selNode(tree = tinyTree, data = dat, 
                 minTip = 4, maxTip = 9,
                 minPr = 0, maxPr = 0.8, all = TRUE))

## provide obj
lse <- TreeSummarizedExperiment(rowTree = tinyTree,
                                assays = list(toyTable))
(out3 <- selNode(obj = lse, all = TRUE))
# All nodes except node 1
(out4 <- selNode(obj = lse, skip = 1, all = TRUE))

}
\author{
Ruizhu Huang
}
