% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetTse.R
\name{subsetTse}
\alias{subsetTse}
\title{subset the TreeSummarizedExperiment object}
\usage{
subsetTse(x, subset_rows, subset_columns, update_rowTree = FALSE,
  update_colTree = FALSE, trim_internal_rowTree = TRUE,
  trim_internal_colTree = TRUE, collapse_singles_rowTree = FALSE,
  collapse_singles_colTree = FALSE, message = FALSE)
}
\arguments{
\item{x}{A TreeSummarizedExperiment object}

\item{subset_rows}{A logical expression indicating rows to keep.}

\item{subset_columns}{A logical expression indicating columns to keep}

\item{update_rowTree}{A logical value, TRUE or FALSE. If TRUE, the row tree
is updated correspondingly.}

\item{update_colTree}{A logical value, TRUE or FALSE. If TRUE, the column
tree is updated correspondingly.}

\item{trim_internal_rowTree}{A logical value, TRUE or FALSE. If TRUE,
internal nodes with chldren removed are dropped from the
\code{rowTree}.}

\item{trim_internal_colTree}{A logical value, TRUE or FALSE. If TRUE,
internal nodes with chldren removed are dropped from the
\code{colTree}.}

\item{collapse_signles_rowTree}{A logical value, TRUE or FALSE. If TRUE,
internal nodes with only one direct child node is removed from the
\code{rowTree}.}

\item{collapse_signles_colTree}{A logical value, TRUE or FALSE. If TRUE,
internal nodes with only one direct child node is removed from the
\code{colTree}.}
}
\value{
A TreeSummarizedExperiment object
}
\description{
\code{subsetTse} is to subset the TreeSummarizedExperiment object. The
corresponding tree could be optionally updated. If the tree is allowed to be
updated, branches with no data in the \code{assays} table will be removed to
create a new tree structure. Users might find leaf nodes that have no
corresponding rows (or columns) in the \code{assays} table are kept in the
\code{rowTree} (or \code{colTree}). That's because at least one of its
ancestor nodes (nodes in the path connecting the leaf and the root) could be
mapped to rows (or columns) in the \code{assays} table.
}
\examples{

data("tinyTree")
set.seed(1)
# the count table
count <- matrix(rpois(100, 50), nrow = 10)
rownames(count) <- c(tinyTree$tip.label)
colnames(count) <- paste("C_", 1:10, sep = "_")

# The sample information
sampC <- data.frame(condition = rep(c("control", "trt"), each = 5),
                    gender = sample(x = 1:2, size = 10, replace = TRUE))
rownames(sampC) <- colnames(count)

# build a TreeSummarizedExperiment object
tse <- TreeSummarizedExperiment(assays = list(count),
                                colData = sampC,
                                rowTree = tinyTree)
rowLinks(tse)
                       
sel <- sample(sample(c(TRUE, FALSE), size = 10, replace = TRUE))         
tse1 <- subsetTse(x = tse, subset_rows = sel, update_rowTree = TRUE )    
rowLinks(tse1) 
}
\author{
Ruizhu Huang
}
