% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{throat}
\alias{throat}
\title{throat data}
\format{An object of class \code{TreeSummarizedExperiment} with 856 rows and 60 columns.}
\usage{
throat
}
\description{
throat data
}
\keyword{datasets}
