% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{throat_v35}
\alias{throat_v35}
\title{a subset data from HMP16SData
A dataset containing 16sRNA counts on body subsite: Throat}
\format{A TreesummarizedExperiment object with 956 OTUs and 153 samples}
\source{
\url{https://gist.github.com/fionarhuang/5f18a93243fc9f901f61451d18eab2aa}
}
\usage{
throat_v35
}
\description{
a subset data from HMP16SData
A dataset containing 16sRNA counts on body subsite: Throat
}
\keyword{datasets}
