% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeclimbR-package.R
\docType{package}
\name{treeclimbR-package}
\alias{treeclimbR}
\alias{treeclimbR-package}
\title{treeclimbR}
\description{
The arrangement of hypotheses in a hierarchical structure appears
in many research fields and often indicates different resolutions at which
data can be viewed. This raises the question of which resolution level
the signal should best be interpreted on. treeclimbR provides a flexible
method to select optimal resolution levels (potentially different levels
in different parts of the tree), rather than cutting the tree at an
arbitrary level. treeclimbR uses a tuning parameter to generate candidate
resolutions and from these selects the optimal one.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csoneson/treeclimbR}
  \item Report bugs at \url{https://github.com/csoneson/treeclimbR/issues}
}

}
\author{
Ruizhu Huang

Charlotte Soneson
}
\keyword{internal}
