% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewBranch.R
\name{viewBranch}
\alias{viewBranch}
\title{view tree and its two zoomed branches}
\usage{
viewBranch(
  tree,
  ann_data = NULL,
  ann_column = NULL,
  ann_color = "black",
  ann_hjust = c(0.2, 0.2),
  ann_vjust = c(0.6, -0.5),
  ann_size = 1,
  hlight_node = NULL,
  hlight_alpha = 0.5,
  hlight_fill = "blue",
  point_node = NULL,
  point_color = "cyan",
  point_size = 3,
  view_node = NULL,
  zoom_node = NULL,
  zoom_scale = 2,
  group_leaf = NULL,
  group_color = "orange",
  nrow = nrow,
  rel_widths = c(0.5, 2, 3),
  branch_length = "none",
  layout = "rectangular",
  edge_size = 1,
  ...
)
}
\arguments{
\item{tree}{A phylo object}

\item{ann_data}{A data frame to annotate nodes on the tree. Default is NULL,
otherwise, it should include at least two columns: one column about nodes
and named as \code{node} and the other column about the annotation value
and could be named freely.}

\item{ann_column}{A character vector to specify columns in \code{ann_data}
that provide values to annotate nodes.}

\item{ann_color}{A character vector to specify the colors to be used for the
annotation. It should have the same length as \code{ann_column}.}

\item{ann_hjust}{A numeric vector. It is to adjust the label locations
horizontally. Please refer to \code{hjust} in \code{ggtree::geom_text2}}

\item{ann_vjust}{A numeric vector. It is to adjust the label locations
vertically. Please refer to \code{vjust} in \code{ggtree::geom_text2}}

\item{ann_size}{A numeric vector. It is to adjust the label size. Please
refer to \code{size} in \code{ggtree::geom_text2}}

\item{hlight_node}{A vector of node number. The branch to be highlight.
Please refer to \code{node} in \code{ggtree::geom_hilight}}

\item{hlight_alpha}{A numeric value. Please refer to \code{alpha} in
\code{ggtree::geom_hilight}}

\item{hlight_fill}{A vector of colors to hilight branches.
Please refer to \code{fill} in \code{ggtree::geom_hilight}}

\item{point_node}{A vector of node number to add points.}

\item{point_color}{A vector of colors to color points.}

\item{point_size}{A numeric value to specify the point size.}

\item{view_node}{A vector of node number. Branches to be shown in subplots.}

\item{zoom_node}{A vector of node number. Branches to be zoomed in. Please
refer to \code{node} in \code{ggtree::scaleClade}}

\item{zoom_scale}{A numeric vector. The zoom scale for \code{zoom_node}.
Please refer to \code{scale} in \code{ggtree::scaleClade}}

\item{group_leaf}{A named list of leaf node number. Please refer to
\code{node} in \code{ggtree::groupOTU}}

\item{group_color}{A named vector provide the color for \code{group_leaf}.
The names should match with the name in \code{group_leaf} and one extra
with name "0". Please see the example.}

\item{nrow}{Please refer to
\code{nrow} of \code{\link[cowplot]{plot_grid}}.}

\item{rel_widths}{The widths of three plots. Please refer to
\code{rel_widths} of \code{\link[cowplot]{plot_grid}}.}

\item{branch_length}{Please refer to the argument \code{branch.length} of
\code{\link[ggtree]{ggtree}}. The default is "none".}

\item{layout}{Please refer to the argument \code{layout} of
\code{\link[ggtree]{ggtree}}. The default is "rectangular".}

\item{edge_size}{Please refer to the argument \code{size} of
\code{\link[ggtree]{ggtree}}. The default is 1.}

\item{...}{More arguments accepted by \code{\link[cowplot]{plot_grid}}.}
}
\value{
a figure
}
\description{
\code{viewBranch} shows a full tree and two zommed branches
}
\examples{
library(TreeSummarizedExperiment)
data("tinyTree")

df <- data.frame(node = printNode(tinyTree, type = "all")$nodeNum,
                 score1 = round(runif(19), 2),
                 score2 = round(runif(19), 2))   
                 
grp <- list(A = 1:3, B = 7:8)                  
viewBranch(tree = tinyTree, ann_data = df,
           ann_column = c("node", "score2"),
           ann_color = c("blue", "red"),
           ann_hjust = c(0.2, 0.2),
           ann_vjust = c(0.6, -0.5),
           ann_size = c(3, 3), 
           view_node = c(16, 14), nrow = 1,
           zoom_node = c(16), zoom_scale = 10, 
           group_leaf = grp,
           group_color = c(A ="orange", B = "blue", "0" = "grey"))
}
