% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewHeatmap.R
\name{viewHeatmap}
\alias{viewHeatmap}
\title{create a heatmap}
\usage{
viewHeatmap(obj, groupCol = "group", scaleColumn = FALSE, ...)
}
\arguments{
\item{obj}{A TreeSummarizedExperiment}

\item{groupCol}{The name of the column that contains the group information of
samples in \code{colData} of \code{obj}.}

\item{scaleColumn}{TRUE or FALSE. If TRUE, the column of assay table is
scaled.}

\item{...}{Arguments that are accepted in
\code{\link[ComplexHeatmap]{Heatmap}}.}
}
\value{
a heatmap
}
\description{
create a heatmap
}
