% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewSim.R
\name{viewSim}
\alias{viewSim}
\title{visualize simulated scenario}
\usage{
viewSim(obj, zoom_scale = 2, legend_position = "left",
  show_leaf = FALSE, ref_value = 1, ...)
}
\arguments{
\item{obj}{The output from \code{simData}}

\item{zoom_scale}{A positive numeric value. If it is above one, branches with
fold change equal to one (non-signal branch) will be zoomed in; If below
one, they will be shrinked. Default is 2}

\item{legend_position}{The legend position. see \code{\link[ggplot2]{theme}}}

\item{show_leaf}{TRUE or FALSE with fold change above or below 1 will be
labelled.}

\item{ref_value}{Default is 1. The fold change of the non-signal branch. (The
slot \code{FC} in \code{metadata} of \strong{obj})}

\item{...}{Additional arguments that are accepted by
\code{\link[ggtree]{ggtree}}.}
}
\value{
a figure
}
\description{
\code{viewSim} is to visualize the output from the function \code{simData}.
}
\examples{
set.seed(1)
y <- matrix(rnbinom(100,size=1,mu=10),nrow=10)
colnames(y) <- paste("S", 1:10, sep = "")
rownames(y) <- tinyTree$tip.label


toy_lse <- TreeSummarizedExperiment(rowTree = tinyTree,
                                    assays = list(y))
res <- parEstimate(obj = toy_lse)

set.seed(1122)
dat1 <- simData(obj = res)
viewSim(obj = dat1, layout = "circular")



}
