% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wScoreTree.R
\name{wScoreTree}
\alias{wScoreTree}
\title{Generate weighted score accounting for the family effect}
\usage{
wScoreTree(tree, scoreData, nodeCol, scoreCol)
}
\arguments{
\item{tree}{A phylo object}

\item{scoreData}{A data frame that includes at least two columns. One column
stores the number of the node, and the other stores the original score of
the corresponding node.}

\item{nodeCol}{The name of the column that stores the number of the node.}

\item{scoreCol}{The name of the column that stores the original score of the
node.}
}
\description{
\code{wScoreTree} takes the tree structure into account when calculating the
score for an internal node. If an internal node A has two children B and C,
\code{(A->B, A->C)}, the new score at node A would be calculated as the
weighted mean of the scores in the whole family (A, B and C). The weights are
based on the number of descendant leaves. For example, if the node B has 2
descendant leaves, and C has 3 descendant leaves, then A would have 5. The
calculation would be \eqn{(Score\_A * 5 + Score\_B*2 +Score\_C*3)/10}. The
generation starts from the leaves and the new generated scores are used to
update those in higher level of the tree until the root is reached.
}
\examples{

library(TreeSummarizedExperiment)
library(ggtree)
library(dplyr)

# tree
data("tinyTree")
ggtree(tinyTree) + geom_text2(aes(label = node))

# score
exScore <- data.frame(nodeNum = 1:19, score = (1:19)/10)

# update the score based on the family provided by the tree
newScore <- wScoreTree(tree = tinyTree, scoreData = exScore,
                      nodeCol = "nodeNum", scoreCol = "score")

# visualize the result
df <- newScore \%>\% 
      rename(node = nodeNum) \%>\%    
      mutate(score = round(score, 3)) \%>\%
      mutate(wScore = round(wScore, 3))

# the original scores are in black texts and the new ones in blue
ggtree(tinyTree) \%<+\% df + 
geom_text2(aes(label = score), hjust = -0.05) +
geom_text2(aes(label = wScore, hjust = -1.2), color = "blue") +
xlim(0, 5)  
}
\author{
Ruizhu HUANG
}
