% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_isConnect.R
\name{isConnect}
\alias{isConnect}
\title{test whether nodes are connected}
\usage{
isConnect(tree, node_a, node_b, connect = "any")
}
\arguments{
\item{tree}{A phylo object}

\item{node_a}{A vector of nodes.}

\item{node_b}{A vector of nodes.}

\item{connect}{It is choosen among "any", "direct", "indirect" and defines
how two (vectors of ) nodes are connected.}
}
\description{
\code{isConnect} tests whether two (vectors of) nodes are connected.
}
\examples{

data("tinyTree")
node_a <- c(4, 18, 19, 2)
node_b <- c(4, 5, 6, 3)

isConnect(tree = tinyTree, node_a = node_a,
          node_b = node_b, connect = "any")


}
\author{
Ruizhu Huang
}
