% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{Simulate different scenarios of abundance change in entities}
\usage{
simData(
  tree = NULL,
  data = NULL,
  obj = NULL,
  scenario = "S1",
  from.A = NULL,
  from.B = NULL,
  minTip.A = 0,
  maxTip.A = Inf,
  minTip.B = 0,
  maxTip.B = Inf,
  minPr.A = 0,
  maxPr.A = 1,
  ratio = 4,
  adjB = NULL,
  pct = 0.6,
  nSam = c(50, 50),
  mu = 10000,
  size = NULL,
  n = 1,
  FUN = sum,
  message = FALSE
)
}
\arguments{
\item{tree}{A phylo object. Only use when \code{obj} is NULL.}

\item{data}{A matrix, representing a table of values, such as count,
collected from real data. It has the entities corresponding to tree leaves
in the row and samples in the column. Only use when \code{obj} is NULL.}

\item{obj}{A TreeSummarizedExperiment object. More details in
\code{\link[TreeSummarizedExperiment]{TreeSummarizedExperiment-class}}. The
\code{tree} (see argument \code{tree}) is stored as \code{rowTree} and
\code{data} (see argument \code{data}) is stored as the first table of
\code{assays}.}

\item{scenario}{\dQuote{S1}, \dQuote{S2}, or \dQuote{S3} (see
\bold{Details}). Default is \dQuote{S1}.}

\item{from.A, from.B}{The branch node labels of branches A and B for which the
signal is swapped. Default, both are NULL. In simulation, we select two
branches (A & B) to have differential abundance under different conditions.
One could specify these two branches or let \code{simData} choose. (Note: If
\code{from.A} is NULL, \code{from.B} is set to NULL).}

\item{minTip.A}{The minimum number of leaves in branch A}

\item{maxTip.A}{The maximum number of leaves in branch A}

\item{minTip.B}{The minimum number of leaves in branch B}

\item{maxTip.B}{The maximum number of leaves in branch B}

\item{minPr.A}{A numeric value selected from 0 to 1. The minimum abundance
proportion of leaves in branch A}

\item{maxPr.A}{A numeric value selected from 0 to 1. The maximum abundance
proportion of leaves in branch A}

\item{ratio}{A numeric value. The proportion ratio of branch B to branch A.
This value is used to select branches(see \bold{Details}). If there are no
branches having exactly this ratio, the pair with the value closest to
\code{ratio} would be selected.}

\item{adjB}{a numeric value selected from 0 and 1 (only for \code{scenario}
is \dQuote{S3}). Default is NULL. If NULL, branch A and the selected part
of branch B swap their proportions. If a numeric value, e.g. 0.1, then the
selected part of branch B decreases to its one tenth proportion and the
decrease in branch B is added to branch A. For example, assume there are
two experimental conditions (C1 & C2), branch A has 10 and branch B has 40
in C1. If adjB is set to 0.1, then in C2 branch B becomes 4 and branch A 46
so that the total proportion stays the same.}

\item{pct}{The percentage of leaves in branch B that have differential
abundance under different conditions (only for scenario \dQuote{S3})}

\item{nSam}{A numeric vector of length 2, containing the sample size for two
different conditions}

\item{mu, size}{The parameters of the Negative Binomial distribution. (see mu
and size in \code{\link[stats:NegBinomial]{rnbinom}}). Parameters used to
generate the library size for each simulated sample. If \code{size} is not
specified, \code{mu} should be a vector of numbers from which the library
size is sampled from with replacement.}

\item{n}{A numeric value to specify how many count tables would be generated
with the same settings. Default is one and one count table would be
obtained at the end. If above one, the output is a list of matrices (count
tables). This is useful, when one needs multiple simulations.}

\item{FUN}{A function to derive the count at each internal node based on its
descendant leaves, e.g. sum, mean. The argument of the function is a
numeric vector with the counts of an internal node's descendant leaves.}

\item{message}{A logical value, TRUE or FALSE. Default is FALSE. If TRUE, the
progress message is printed out.}
}
\value{
a TreeSummarizedExperiment object.
 \itemize{
 \item \strong{assays} a list of count table or a count table. Entities on
 the row and samples in the column. Each row could be mapped to a node of the
 tree.
 \item \strong{rowData} the annotation data for the rows of tables in
 \code{assays}.
 \item \strong{colData} the annotation data for the columns of tables in
 \code{assays}.
 \item \strong{rowTree} the tree structure of entities.
 \item \strong{rowLinks} the link between rows of \code{assays} and nodes on
 the tree.
 \item \strong{metadata} more details about the simulation.
   \itemize{
   \item \strong{FC} the fold change of entities correspondint to the tree
   leaves.
   \item \strong{Branch} the information about two selected branches.
     \itemize{
       \item \strong{A} the branch node label (or number) of branch A
       \item \strong{B} the branch node label (or number) of branch B
       \item \strong{ratio} the count proportion ratio of branch B to branch
       A
       \item \strong{A_tips} the number of leaves on branch A
       \item \strong{B_tips} the number of leaves on branch B
       \item \strong{A_prop} the count proportion of branch A
       (a value between 0 and 1)
       \item \strong{B_prop} the count proportion of branch B
       (a value between 0 and 1)
     }
     }}
}
\description{
\code{simData} simulates different abundance patterns for entities under
different conditions. These entities have their corresponding nodes on a
tree. More details about the simulated patterns could be found in the
vignette via \code{browseVignettes("treeclimbR")}.
}
\details{
\code{simData} simulates a count table for entities which are
  corresponding to the nodes of a tree. The entities are in rows and the
  samples from different groups or conditions are in columns. The library
  size of each sample is sampled from a Negative Binomial distribution with
  mean and size specified by the arguments \code{mu} and \code{size}. The
  counts of entities, that are mapped ot the leaf nodes, in a same sample are
  assumed to follow a Dirichlet-Multinomial distribution. The parameters for
  the Dirichlet-Multinomial distribution are estimated from a real data set
  specified by the argument \code{data} via the function \code{dirmult} (see
  \code{\link[dirmult]{dirmult}}). To generate different abundance patterns
  under different conditions, we provide three different scenarios,
  \dQuote{S1}, \dQuote{S2}, and \dQuote{S3} (specified via \code{scenario}).
  Our vignette provides figures to explain these three scenarios (try
  \code{browseVignettes("treeAGG")}). \itemize{ \item S1: two branches are
  selected to swap their proportions, and leaves on the same branch have the
  same fold change. \item S2: two branches are selected to swap their
  proportions. Leaves in the same branch have different fold changes but same
  direction (either increase or decrease). \item S3: two branches are
  selected. One branch has its proportion swapped with the proportion of some
  leaves from the other branch.}
}
\examples{

set.seed(1)
y <- matrix(rnbinom(100,size=1,mu=10),nrow=10)
colnames(y) <- paste("S", 1:10, sep = "")
rownames(y) <- tinyTree$tip.label


toy_lse <- TreeSummarizedExperiment(rowTree = tinyTree,
                                    assays = list(y))
res <- parEstimate(obj = toy_lse)

set.seed(1122)
dat1 <- simData(obj = res, ratio = 2, scenario = "S1", pct = 0.5)



}
\author{
Ruizhu Huang
}
